# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance RESTful backend API for the Timesheet Automation Portal, built with Go (Gin Engine) and PostgreSQL.\n- Automated Timesheet Generator: Multi-company Excel (.xlsx) and PDF generation (MII, SDD, Adidata, NTT) with day-fraction calculations, dynamic day trimming, formula preservation, and async SMTP delivery.\n- Dynamic Template Engine: Dedicated programmatic template builders alongside visual cell-mapping coordinate parsers integrated with Handsontable.\n- Dual Authentication \u0026 Security: JWT sessions with NIST SP 800-63B compliant Argon2id password hashing and WebAuthn FIDO2 Passkeys (biometric \u0026 hardware-bound).\n- Role-Based Access Control (RBAC): Admin and User role segregation, centralized account administration, and profile change approval workflows.\n- Web Push Notifications: Daily reminder scheduler pinned to 17:00 WIB (Asia/Jakarta) notifying active users via VAPID.\n- 3NF Relational Master Data: Normalized PostgreSQL database schema (Companies, Departments, Projects, Holidays, Activity Statuses) with foreign keys and automated SQL migrations.\n- API Versioning: Standardized RESTful endpoints routed under the `/api/v1/*` namespace.",
        "title": "Timesheet Automation Portal API",
        "termsOfService": "https://github.com/naufalzaid17/timesheet-generator",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
...
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/templates": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ingests an admin-uploaded .xlsx workbook template for timesheet generation.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Upload template workbook (Admin)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel .xlsx template file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Template display name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Template description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Associated company code or name",
                        "name": "company",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Target worksheet name",
                        "name": "sheet_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Set as default template (true/false)",
                        "name": "is_default",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Template"
                        }
                    },
                    "400": {
                        "description": "Invalid file or parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Admin only",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/templates/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a template and its mappings (admin only).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete template (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Admin only",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/templates/{id}/default": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks a template as the default used for generation when a user doesn't pick one (admin only).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Set default template (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Admin only",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all uploaded timesheet templates with their cell mappings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "List templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Template"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/templates/{id}/grid": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the parsed 2-D cell grid and merged cells for Handsontable rendering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "Get template grid layout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.SheetLayout"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CellMapping": {
            "type": "object",
            "properties": {
                "cell_ref": {
                    "description": "For ScopeCell: absolute address, e.g. \"C4\".",
                    "type": "string"
                },
                "column": {
                    "description": "For ScopeDailyColumn: the column letter (e.g. \"K\") and the row where the\nfirst day of the month is written.",
                    "type": "string"
                },
                "field": {
                    "$ref": "#/definitions/models.MappingFieldType"
                },
                "fillable": {
                    "description": "Fillable marks whether users may edit this field in the monthly grid.",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "scope": {
                    "$ref": "#/definitions/models.MappingScope"
                },
                "start_row": {
                    "type": "integer"
                },
                "template_id": {
                    "type": "integer"
                }
            }
        },
        "models.Company": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "\"mii\", \"sdd\", \"adidata\", \"ntt\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "departments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Department"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Project"
                    }
                },
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Template"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Department": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"WCSD\", \"DEV-01\"",
                    "type": "string"
                },
                "company": {
                    "$ref": "#/definitions/models.Company"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "division": {
                    "description": "e.g. \"Wholesale Digital Delivery\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "description": "e.g. \"Wholesale Channel and Service Delivery\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request or unauthorized"
                }
            }
        },
...
        "models.MappingFieldType": {
            "type": "string",
            "enum": [
                "date",
                "time_in",
                "time_out",
                "status",
                "activity",
                "project_name",
                "project_id",
                "app_impacted",
                "total_hour",
                "division",
                "department",
                "sub_department",
                "aip_fitur",
                "meta_name",
                "meta_mii_id",
                "meta_division",
                "meta_site",
                "meta_month",
                "meta_year"
            ],
            "x-enum-comments": {
                "FieldDepartment": "per-day departement column",
                "FieldDivision": "per-day divisi column",
                "FieldTotalHour": "End - Start (computed)"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "End - Start (computed)",
                "per-day divisi column",
                "per-day departement column",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "FieldDate",
                "FieldTimeIn",
                "FieldTimeOut",
                "FieldStatus",
                "FieldActivity",
                "FieldProjectName",
                "FieldProjectID",
                "FieldAppImpacted",
                "FieldTotalHour",
                "FieldDivision",
                "FieldDepartment",
                "FieldSubDept",
                "FieldAIPFitur",
                "FieldMetaName",
                "FieldMetaMiiID",
                "FieldMetaDivision",
                "FieldMetaSite",
                "FieldMetaMonth",
                "FieldMetaYear"
            ]
        },
        "models.MappingScope": {
            "type": "string",
            "enum": [
                "cell",
                "daily_column"
            ],
            "x-enum-varnames": [
                "ScopeCell",
                "ScopeDailyColumn"
            ]
        },
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
...
        "models.Project": {
            "type": "object",
            "properties": {
                "app_impacted": {
                    "description": "e.g. \"BNI Direct Cash\"",
                    "type": "string"
                },
                "code": {
                    "description": "e.g. \"P24015\"",
                    "type": "string"
                },
                "company": {
                    "$ref": "#/definitions/models.Company"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "description": "e.g. \"BNI Direct\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Role": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser"
            ]
        },
...
        "models.Template": {
            "type": "object",
            "properties": {
                "builtin": {
                    "description": "Builtin, when set (e.g. \"bni_dev\"), marks a first-class bundled template\nwhose fixed layout is rendered by a dedicated strict-typed generator rather\nthan the generic cell-mapping engine.",
                    "type": "string"
                },
                "cell_mappings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CellMapping"
                    }
                },
                "company": {
                    "type": "string"
                },
                "company_id": {
                    "type": "integer"
                },
                "company_rel": {
                    "$ref": "#/definitions/models.Company"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/models.User"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sheet_name": {
                    "description": "SheetName is the worksheet the mapping applies to.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "company_id": {
                    "type": "integer"
                },
                "company_rel": {
                    "$ref": "#/definitions/models.Company"
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "department_id": {
                    "type": "integer"
                },
                "department_rel": {
                    "$ref": "#/definitions/models.Department"
                },
                "division": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employee_id": {
                    "description": "NPP or MII ID",
                    "type": "string"
                },
                "group_name": {
                    "description": "Kelompok (SDD)",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "mii_id": {
                    "type": "string"
                },
                "name": {
                    "description": "Profile fields (the \"approved\" / live values).",
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "site": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "services.MergeRegion": {
            "type": "object",
            "properties": {
                "col": {
                    "type": "integer"
                },
                "colspan": {
                    "type": "integer"
                },
                "row": {
                    "type": "integer"
                },
                "rowspan": {
                    "type": "integer"
                }
            }
        },
        "services.SheetLayout": {
            "type": "object",
            "properties": {
                "col_widths": {
                    "description": "pixels, per column",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "grid": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "merges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.MergeRegion"
                    }
                },
                "sheet_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with format \"Bearer {token}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}